requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
aiomysql==0.2.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
discord.py==2.3.2
idna==3.4
instaloader==4.10
multidict==6.0.4
oauthlib==3.2.2
requests==2.31.0
requests-oauthlib==1.3.1
soupsieve==2.4.1
urllib3==2.0.4
yarl==1.9.2
pypresence==4.3.0
lxml==4.9.3
Pillow==9.5.0
aiosqlite==0.17.0
mtranslate==1.8
mysql-connector-python==8.1.0
Flask==2.3.3
spotipy==2.23.0
pytz==2023.3